+++ mcp9808/mcp9808.h
@@ -1,7 +1,7 @@
 #ifndef MCP9808_H
 #define MCP9808_H
 
-#include <i2c_funcs.h>
+#include <libi2c.h>
 #include <stdint.h>
 #include <stdbool.h>
 
@@ -12,7 +12,7 @@ typedef struct mcp9808
 
 } MCP9808;
 
-bool mcp9808_init(MCP9808 *mcp9808, int channel, int addr);
+bool mcp9808_init(MCP9808 *mcp9808, int channel, uint8_t addr);
 bool mcp9808_read(MCP9808 *mcp9808, float *result);
 
 #endif // MCP9808_H
